LPS-56048 Move asset components into module
