+++ modules/apps/message-boards/message-boards-web/src/com/liferay/message/boards/web/asset/MBMessageAssetRendererFactory.java
@@ -12,16 +12,16 @@
  * details.
  */
 
-package com.liferay.portlet.messageboards.asset;
+package com.liferay.message.boards.web.asset;
 
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.portlet.LiferayPortletResponse;
 import com.liferay.portal.kernel.portlet.LiferayPortletURL;
-import com.liferay.portal.kernel.spring.osgi.OSGiBeanProperties;
 import com.liferay.portal.security.permission.PermissionChecker;
 import com.liferay.portal.theme.ThemeDisplay;
 import com.liferay.portal.util.PortletKeys;
 import com.liferay.portlet.asset.model.AssetRenderer;
+import com.liferay.portlet.asset.model.AssetRendererFactory;
 import com.liferay.portlet.asset.model.BaseAssetRendererFactory;
 import com.liferay.portlet.messageboards.model.MBMessage;
 import com.liferay.portlet.messageboards.service.MBMessageLocalServiceUtil;
@@ -32,16 +32,21 @@ import javax.portlet.PortletURL;
 import javax.portlet.WindowState;
 import javax.portlet.WindowStateException;
 
+import org.osgi.service.component.annotations.Component;
+
 /**
  * @author Julio Camarero
  * @author Juan Fernández
  * @author Raymond Augé
  * @author Sergio González
  */
-@OSGiBeanProperties(
+@Component(
+	immediate = true,
 	property = {
+		"javax.portlet.name=" + PortletKeys.MESSAGE_BOARDS,
 		"search.asset.type=com.liferay.portlet.messageboards.model.MBMessage"
-	}
+	},
+	service = AssetRendererFactory.class
 )
 public class MBMessageAssetRendererFactory extends BaseAssetRendererFactory {
