+++ portal-test-internal/src/com/liferay/portal/service/test/PortalRegisterTestUtil.java
@@ -26,8 +26,6 @@ import com.liferay.portlet.documentlibrary.asset.DLFileEntryAssetRendererFactory
 import com.liferay.portlet.documentlibrary.asset.DLFolderAssetRendererFactory;
 import com.liferay.portlet.documentlibrary.util.DLFileEntryIndexer;
 import com.liferay.portlet.documentlibrary.util.DLFolderIndexer;
-import com.liferay.portlet.messageboards.asset.MBCategoryAssetRendererFactory;
-import com.liferay.portlet.messageboards.asset.MBMessageAssetRendererFactory;
 import com.liferay.portlet.messageboards.trash.MBCategoryTrashHandler;
 import com.liferay.portlet.messageboards.trash.MBMessageTrashHandler;
 import com.liferay.portlet.messageboards.trash.MBThreadTrashHandler;
@@ -113,9 +111,7 @@ public class PortalRegisterTestUtil {
 
 	private static final Class<?>[] _ASSET_RENDERER_FACTORY_CLASSES = {
 		DLFileEntryAssetRendererFactory.class,
-		DLFolderAssetRendererFactory.class,
-		MBCategoryAssetRendererFactory.class,
-		MBMessageAssetRendererFactory.class, UserAssetRendererFactory.class
+		DLFolderAssetRendererFactory.class, UserAssetRendererFactory.class
 	};
 
 	private static boolean _assetRendererFactoriesRegistered;
